% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_FEAT_SELECT_SITE.R
\name{calc_fit_site}
\alias{calc_fit_site}
\title{Function for calculating 'fitness' of subset of sites based on cumulative kolmogorov smirnov D statistic}
\usage{
calc_fit_site(sample, pop_df)
}
\arguments{
\item{sample}{Integer vector representing subset of sites (rows) in the candidate site data.frame.}

\item{pop_df}{Data.frame with same covariates (columns) as the candidate site data.frame, but assumed to represent the population distribution.}
}
\value{
Cumulative kolmogorov smirnov D statistic for each covaraite in the sample of sites with respect to the population of sites (i.e, all possible locations)
}
\description{
Function for calculating 'fitness' of subset of sites based on cumulative kolmogorov smirnov D statistic
}
