% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_FEAT_SELECT.R
\name{cross_over}
\alias{cross_over}
\title{Function for crossing two parent models.}
\usage{
cross_over(PARENTS, MUTE_RATE, MAX_TERMS, Y_NAME, X_NAMES)
}
\arguments{
\item{PARENTS}{A list of parent models.}

\item{MUTE_RATE}{A rate of mutation,0-1.}

\item{MAX_TERMS}{Maximum number of terms to include in a model, does not account for interactions.}

\item{Y_NAME}{String name of response variable in data frame.}

\item{X_NAMES}{Character vector of possible predictors present in data frame.}
}
\value{
A model with genetic material 'input features' from parents, as a formula.
}
\description{
This function randomly combines two models, with
somewhere between 1 and MAX_TERMS terms. Models are
subjected to mutation controlled by a mutation rate.
}
