% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_FEAT_SELECT_SITE.R
\name{crossover_site}
\alias{crossover_site}
\title{Function for randomly selecting site subset from all subsets with fitness below a specified percentile.}
\usage{
crossover_site(parents, mute_rate, samp_df)
}
\arguments{
\item{parents}{List of two parent subsets of sites.}

\item{mute_rate}{Rate of mutation 0-1.}

\item{samp_df}{Data frame containing all covarites of interest as columns, and all potential sites as rows.}
}
\value{
A new subset of sites that is the random combination of the two provided parent subsets, with mutation.
}
\description{
Function for randomly selecting site subset from all subsets with fitness below a specified percentile.
}
