% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_FEAT_SELECT_SITE.R
\name{evolve_site}
\alias{evolve_site}
\title{Function for evolving a subset of site locations from a population of canidate site locations for observational regression study design.}
\usage{
evolve_site(
  N,
  num_sites,
  canidate_df,
  population_df,
  generations,
  prctl_thresh,
  mute_rate
)
}
\arguments{
\item{N}{Total number of random subsets of sites to initialize,i.e., population size.}

\item{num_sites}{Maximum number of sites to include in a subset, i.e., total number of sample sites.}

\item{canidate_df}{Data frame containing all covarites of interest as columns, and all canidate sites locations as rows.}

\item{population_df}{Data frame containing all covarites of interest as columns, and a discrete approximation of all possible sites locations in the area of interest as rows.}

\item{generations}{Total number of generations to iterate.}

\item{prctl_thresh}{Fitness percentile for which any subset of sites with a fitness score below this will be canidates for cross-over.}

\item{mute_rate}{Rate of mutation during cross-over (0-1).}
}
\value{
A new subset of sites that is the random combination of the two provided parent subsets, with mutation.
}
\description{
Function for evolving a subset of site locations from a population of canidate site locations for observational regression study design.
}
