% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_FEAT_SELECT.R
\name{init_pop}
\alias{init_pop}
\title{Initializes a population of random models.}
\usage{
init_pop(Y_NAME, X_NAMES, MAX_TERMS, N)
}
\arguments{
\item{Y_NAME}{String, name of response variable in data frame.}

\item{X_NAMES}{Character vector of possible predictor names present in data frame.}

\item{MAX_TERMS}{Maximum number of terms to include in a model, does not account for interactions.}

\item{N}{Total number of random individual models to initialize,i.e., population size.}
}
\value{
A list 'population' of random models as formulas.
}
\description{
This function initializes a population of random models
provided predictors and a response variable as character vectors.
}
