% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_FEAT_SELECT_SITE.R
\name{init_pop_site}
\alias{init_pop_site}
\title{Initializes a population of random subset of sites (rows).}
\usage{
init_pop_site(N, num_sites, samp_df)
}
\arguments{
\item{N}{Total number of random subsets of sites to initialize,i.e., population size.}

\item{num_sites}{Maximum number of sites to include in a subset, i.e., total number of sample sites.}

\item{samp_df}{Data frame containing all covarites of interest as columns, and all potential sites as rows.}
}
\value{
A list 'population' of random subsets of sites (rows) from the samp_df data.frame.
}
\description{
This function initializes a population of random sites
provided a data frame of possible site locations with
covariates as columns.
}
