% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_FEAT_SELECT.R
\name{select_parents}
\alias{select_parents}
\title{Function for selecting two parents from a population}
\usage{
select_parents(POP, FIT_VECT, PRCTL, MAXIMIZE)
}
\arguments{
\item{POP}{A population list}

\item{FIT_VECT}{A vector of fitness scores corresponding to the population list.}

\item{PRCTL}{The percentile at which the fitness threshold is defined for selection.}

\item{MAXIMIZE}{Boolean, is fitness being maximized, or minimized 'e.g., AIC'}
}
\value{
Output fitness of user defined fitness function.
}
\description{
This function randomly selects to models 'parents' from a population
based on a fitness threshold, i.e., from the pool of models
with a fitness score greater than 'or less than' some percentile.
}
