% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA_FEAT_SELECT_SITE.R
\name{select_site}
\alias{select_site}
\title{Function for randomly selecting site subset from all subsets with fitness below a specified percentile.}
\usage{
select_site(fit_v, prctl)
}
\arguments{
\item{fit_v}{Vector of fitness scores.}

\item{prctl}{Fitness percentile for which any subset of sites with a fitness score below this will be canidates for cross-over.}
}
\value{
Cumulative kolmogorov smirnov D statistic for each covaraite in the sample of sites with respect to the population of sites (i.e, all possible locations)
}
\description{
Function for randomly selecting site subset from all subsets with fitness below a specified percentile.
}
